virtio/virtio_example_driver.c : Introducing the virtio example device driver.
At this stage this is a replica of the PCI example device driver with
all irrelevant part in comment.

For now the only thing the driver does is to enable the device and
print some messages to console when the driver is loaded and removed.

virtio/Makefile : Introducing a simple makefile to build the driver.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>